Replace `ricochet2200/go-disk-usage` for FreeBSD compatibility
`ricochet2200/go-disk-usage` is old, unmaintained library. It uses
deprecated `syscall` package, and it's incompatible with FreeBSD.

I created a fork, that solves both mentioned issues, is properly
versioned, adds Go module support, and I'm willing to maintain it,
if necessary.

This PR makes it possible to compile and run `post` on FreeBSD.
It's also the only code-level change required to use `go-spacemesh`
on FreeBSD.